+++ web/src/views/login.vue
@@ -11,7 +11,14 @@
           <el-input v-model="form.username" placeholder="请输入用户名" size="large" :prefix-icon="User" />
         </el-form-item>
         <el-form-item prop="password">
-          <el-input v-model="form.password" type="password" placeholder="请输入密码" size="large" :prefix-icon="Lock" show-password />
+          <el-input
+            v-model="form.password"
+            type="password"
+            placeholder="请输入密码"
+            size="large"
+            :prefix-icon="Lock"
+            show-password
+          />
         </el-form-item>
         <el-form-item>
           <el-button type="primary" size="large" :loading="loading" native-type="submit" class="login-btn">